+++ src/extension.js
@@ -14,12 +14,15 @@ function activate(context) {
 
     // This function add the selected folder to workspace (VSC Workspace).
     initAddFolderToWorkspace(context);
+
+    // This function removes the selected folder from workspace (VSC Workspace).
+    initRemoveFolderFromWorkspace(context);
 }
 
 function initAddFolderToWorkspace(context) {
 
     const addFolderToWorkspaceId = 'addFolderToWorkspace';
-    context.subscriptions.push(vscode.commands.registerCommand(addFolderToWorkspaceId, () => {
+    context.subscriptions.push(vscode.commands.registerCommand(addFolderToWorkspaceId, async () => {
 
         let workspaceDirectories = [];
         let config = vscode.workspace.getConfiguration('addFolderToWorkspace');
@@ -48,19 +51,88 @@ function initAddFolderToWorkspace(context) {
         }
 
         // Open QuickPick and add selected Folder (Directory to VSC Workspace).
-        vscode.window.showQuickPick(workspaceDirectories, {
+        let workspaces = await vscode.window.showQuickPick(workspaceDirectories, {
+            title: 'AddFolderToWorkspace',
             placeHolder: 'AddFolderToWorkspace: Select a folder...',
-        }).then((Selected) => {
-            if (!Selected) return;
+            canPickMany: true,
+        })
+
+        if (!workspaces) return;
+
+        let workspaceURIs = [];
+        for await (const workspace of workspaces) {
 
             // Get URI of selected directory.
-            let URI = vscode.Uri.file(Selected);
+            let URI = vscode.Uri.file(workspace);
             if (!URI) return;
+            workspaceURIs.push({ uri: URI });
+        }
 
-            // Add selected Folder to Workspace.
-            vscode.workspace.updateWorkspaceFolders(vscode.workspace.workspaceFolders ? vscode.workspace.workspaceFolders.length : 0, null, { uri: URI });
-        });
+        // Add selected Folder to Workspace.
+        await updateWorkspaceAndWait(0, null, workspaceURIs);
+
     }))
+}
+
+function initRemoveFolderFromWorkspace(context) {
+
+    const removeFolderFromWorkspaceId = 'removeFolderFromWorkspace';
+    context.subscriptions.push(vscode.commands.registerCommand(removeFolderFromWorkspaceId, async () => {
+
+        // Check all current workspace folders.
+        let workspaceFolders = [];
+        vscode.workspace.workspaceFolders.sort().forEach(function (workspaceFolder) {
+            workspaceFolders.push(workspaceFolder.name)
+        })
+
+        if (!workspaceFolders.length) return;
+
+        // Create showQuickPick 'RemoveFolderFromWorkspace' selection.
+        let workspaces = await vscode.window.showQuickPick(workspaceFolders, {
+            title: 'RemoveFolderFromWorkspace',
+            placeHolder: 'RemoveFolderFromWorkspace: Select workspaces to be removed...',
+            canPickMany: true,
+        });
+
+        if (!workspaces) return;
+
+        let removeIndexes = [];
+
+        // Sort and reverse selected 'remove' Folder from Workspace.
+        vscode.workspace.workspaceFolders.sort().forEach(function (workspaceFolder) {
+            // workspaceFolders.push(workspaceFolder.name)
+            let removeWorkspace = workspaces.includes(workspaceFolder.name);
+
+            if (removeWorkspace){
+                removeIndexes.push(workspaceFolder.index)
+            }
+        })
+
+        // Remove selected Folder from Workspace.
+        for await (const removeIndex of removeIndexes.reverse()) {
+            await updateWorkspaceAndWait(removeIndex, 1, []);
+        }
+    }))
+}
+
+function updateWorkspaceAndWait(start, deleteCount, workspaceFoldersToAdd) {
+    const success = vscode.workspace.updateWorkspaceFolders(start, deleteCount, ...workspaceFoldersToAdd)
+
+    if (success) {
+        const disps = []
+        return new Promise(resolve => {
+
+            // Note: it is not valid to call updateWorkspaceFolders() multiple times
+            // without waiting for the onDidChangeWorkspaceFolders() to fire.
+            // So we have to always wait in case we want to add or remove multiple folders.
+            vscode.workspace.onDidChangeWorkspaceFolders(() => {
+                resolve();
+            }, null, disps);
+
+        }).finally(() => disps.forEach(disp => disp.dispose()));
+    } else {
+        return Promise.reject(new Error("Failed to update workspace"))
+    }
 }
 
 // This method is called when your extension is deactivated.