+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to the "AddFolderToWorkspace" extension will be documented in this file.
 
+## [1.1.0]
+
+### AddFolderToWorkspace
+
+This Function provides a searchable list of folders (Workspaces) that can be added **simultaneous** to the current VSC Workspace. All configured folders will be displayed.
+
+**Shortcut:** ```strg + alt + k, p```<br>
+**Command:**  ```AddFolderToWorkspace: Add Folder to Workspace.```
+
+### RemoveFolderFromWorkspace
+
+This Function provides a searchable list of folders (Workspaces) that can be removed **simultaneous** from the current VSC Workspace. All current open folders are displayed.
+
+**Shortcut:** ```strg + alt + k, shift + p```<br>
+**Command:**  ```AddFolderToWorkspace: Remove Folder from Workspace.```
+
 ## [1.0.2]
 
 ### Maintenance